listener: extract ws_outbound_task into web_gateway/ws_session.rs
First slice of de-nesting spawn_web_gateway: the outbound WS writer
(broadcast + direct responses + personalized input-authority frames)
becomes a named async fn taking its seven captures as parameters; the
call site spawns ws_outbound_task(...) where the 118-line closure was.
The _outbound-suffixed clone names drop their suffix inside the fn, and
three setup bindings lose a now-unneeded mut. web_gateway tests 405/405.
